Decoupled badge and notification context
