+++ translate/src/modules/editor/hooks/useSendTranslation.ts
@@ -11,6 +11,7 @@ import { EntityView } from '~/context/EntityView';
 import { FailedChecksData } from '~/context/FailedChecksData';
 import { Locale } from '~/context/Locale';
 import { Location } from '~/context/Location';
+import { ShowNotification } from '~/context/Notification';
 import { UnsavedActions } from '~/context/UnsavedChanges';
 import { updateEntityTranslation } from '~/modules/entities/actions';
 import { usePushNextTranslatable } from '~/modules/entities/hooks';
@@ -22,7 +23,7 @@ import { updateResource } from '~/modules/resource/actions';
 import { updateStats } from '~/modules/stats/actions';
 import { useAppDispatch, useAppSelector } from '~/hooks';
 import { serializeEntry, getPlainMessage } from '~/utils/message';
-import { ShowBadgeTooltip, ShowNotification } from '~/context/Notification';
+import { ShowBadgeTooltip } from '~/context/BadgeTooltip';
 
 /**
  * Return a function to send a translation to the server.
