+++ translate/src/modules/editor/hooks/useUpdateTranslationStatus.ts
@@ -7,7 +7,8 @@ import { EntityView } from '~/context/EntityView';
 import { FailedChecksData } from '~/context/FailedChecksData';
 import { HistoryData } from '~/context/HistoryData';
 import { Location } from '~/context/Location';
-import { ShowBadgeTooltip, ShowNotification } from '~/context/Notification';
+import { ShowNotification } from '~/context/Notification';
+import { ShowBadgeTooltip } from '~/context/BadgeTooltip';
 import { updateEntityTranslation } from '~/modules/entities/actions';
 import { usePushNextTranslatable } from '~/modules/entities/hooks';
 import {
